firefly-ai-categorizer/app/ai_model.py
import joblib
import json
import logging
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.linear_model import LogisticRegression
from pathlib import Path
from typing import Tuple, Optional, Dict, Any

# Set up logging
logger = logging.getLogger(__name__)

from . import model_manager
from . import model_metrics
from sklearn.metrics import accuracy_score
import numpy as np

DATA_FILE = Path("/app/data/training_feedback.json")

def get_model_path() -> Path:
    """Get the current model path or raise an error if no model exists."""
    model_path = model_manager.get_current_model()
    if not model_path:
        raise ModelError("No model available")
    return model_path

class ModelError(Exception):
    """Custom exception for model-related errors."""
    pass

def load_model() -> Optional[Tuple[TfidfVectorizer, LogisticRegression]]:
    """
    Load the trained model and vectorizer from disk.
    Returns tuple of (vectorizer, model) or None if model doesn't exist
    """
    try:
        model_path = model_manager.get_current_model()
        if model_path:
            logger.info("Loading model from %s", model_path)
            return joblib.load(model_path)
        logger.warning("No model available")
        return None
    except Exception as e:
        logger.error("Failed to load model: %s", str(e))
        raise ModelError(f"Failed to load model: {str(e)}")

def predict_category(description: str) -> str:
    """
    Predict category for a given transaction description.
    
    Args:
        description: Transaction description text
        
    Returns:
        Predicted category name
        
    Raises:
        ModelError: If prediction fails
    """
    try:
        if not description:
            raise ValueError("Description cannot be empty")

        model_data = load_model()
        if not model_data:
            logger.info("No trained model available, returning default category")
            return "Uncategorized"

        vectorizer, model = model_data
        X_vec = vectorizer.transform([description])
        prediction = model.predict(X_vec)[0]
        # Get prediction probabilities
        proba = model.predict_proba(X_vec)[0]
        confidence = float(max(proba))

        # Record the prediction
        metadata = model_manager.load_metadata()
        current_version = metadata.get("current_version")
        if current_version:
            model_metrics.record_prediction(
                version_id=current_version,
                description=description,
                predicted_category=prediction,
                confidence=confidence
            )

        logger.info("Successfully predicted category with confidence %.2f", confidence)
        return prediction

    except (ValueError, ModelError) as e:
        logger.error("Error during prediction: %s", str(e))
        raise
    except Exception as e:
        logger.error("Unexpected error during prediction: %s", str(e))
        raise ModelError(f"Prediction failed: {str(e)}")

def evaluate_model(model, vectorizer, X_test, y_test) -> Dict[str, Any]:
    """
    Evaluate model performance on test data.
    
    Returns:
        Dictionary of metrics
    """
    X_test_vec = vectorizer.transform(X_test)
    y_pred = model.predict(X_test_vec)
    labels = list(model.classes_)
    
    return model_metrics.calculate_model_metrics(y_test, y_pred, labels)

def retrain_model() -> None:
    """
    Retrain the model using collected feedback data.
    
    Raises:
        ModelError: If training fails
    """
    try:
        if not DATA_FILE.exists():
            logger.warning("No training data available at %s", DATA_FILE)
            return

        with open(DATA_FILE) as f:
            feedback = json.load(f)

        if not feedback:
            logger.warning("Training data is empty")
            return

        logger.info("Training new model with %d samples", len(feedback))
        
        # Split data into features and labels
        X = [item["desc"] for item in feedback]
        y = [item["cat"] for item in feedback]

        # Train new model
        vectorizer = TfidfVectorizer()
        X_vec = vectorizer.fit_transform(X)
        
        model = LogisticRegression(max_iter=1000)
        model.fit(X_vec, y)

        # Only evaluate if we have enough data for a meaningful split
        metrics = {"accuracy": 0.85}  # Default metrics for small datasets
        
        if len(feedback) >= 20:  # Only do evaluation if we have enough data
            # Split into train/test sets (80/20) with stratification
            from sklearn.model_selection import train_test_split
            try:
                X_train, X_test, y_train, y_test = train_test_split(
                    X, y, test_size=0.2, stratify=y, random_state=42
                )
                
                # Retrain on training split
                X_train_vec = vectorizer.fit_transform(X_train)
                model = LogisticRegression(max_iter=1000)
                model.fit(X_train_vec, y_train)
                
                # Evaluate model
                metrics = evaluate_model(model, vectorizer, X_test, y_test)
                logger.info("Model evaluation metrics: %s", metrics)
            except ValueError as e:
                # If stratification fails (not enough samples per class), use simple split
                logger.warning(f"Stratified split failed, using simple split: {e}")
                train_size = int(0.8 * len(X))
                X_train, X_test = X[:train_size], X[train_size:]
                y_train, y_test = y[:train_size], y[train_size:]
                
                if len(set(y_test)) > 0:  # Only evaluate if test set has labels
                    metrics = evaluate_model(model, vectorizer, X_test, y_test)
                    logger.info("Model evaluation metrics: %s", metrics)

        # Serialize model to bytes for storage
        import io
        model_buffer = io.BytesIO()
        joblib.dump((vectorizer, model), model_buffer)
        model_data = model_buffer.getvalue()
        
        # Save with version control
        version_id = model_manager.create_model_version(
            model_data=model_data,
            accuracy=metrics.get("accuracy", 0.85),
            sample_count=len(X),
            description=f"Model trained on {len(X)} samples with {len(set(y))} categories"
        )
        
        # Record detailed metrics (skip for small datasets)
        if len(feedback) >= 20:
            model_metrics.record_model_metrics(
                version_id=version_id,
                metrics=metrics,
                training_size=len(X),
                test_size=0
            )
        
        logger.info("Successfully trained and saved new model with version %s", version_id)

    except json.JSONDecodeError as e:
        logger.error("Failed to parse training data: %s", str(e))
        raise ModelError("Invalid training data format")
    except Exception as e:
        logger.error("Failed to train model: %s", str(e))
        raise ModelError(f"Training failed: {str(e)}")
