firefly-ai-categorizer/requirements.txt
fastapi
uvicorn
httpx
scikit-learn
numpy
pandas
joblib
python-dotenv
requests